pyproject.toml
[project]
name = "ringgem"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Taylor Monacelli", email = "taylormonacelli@gmail.com" }
]
dependencies = [
    "ruamel.yaml>=0.18.6",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/ringgem"]
